fix: validate with envaction when its from bff, validate without when its from api/public
