+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/use_cases/missions/CreateOrUpdateMissionWithActionsAndAttachedReporting.kt
@@ -2,13 +2,14 @@
 
 package fr.gouv.cacem.monitorenv.domain.use_cases.missions
 
-// import fr.gouv.cacem.monitorenv.domain.validators.mission.MissionValidator
 import fr.gouv.cacem.monitorenv.config.UseCase
 import fr.gouv.cacem.monitorenv.domain.entities.mission.MissionEntity
 import fr.gouv.cacem.monitorenv.domain.exceptions.ReportingAlreadyAttachedException
 import fr.gouv.cacem.monitorenv.domain.repositories.IReportingRepository
 import fr.gouv.cacem.monitorenv.domain.use_cases.missions.dtos.EnvActionAttachedToReportingIds
 import fr.gouv.cacem.monitorenv.domain.use_cases.missions.dtos.MissionDetailsDTO
+import fr.gouv.cacem.monitorenv.domain.validators.UseCaseValidation
+import fr.gouv.cacem.monitorenv.domain.validators.mission.MissionWithEnvActionsValidator
 import org.slf4j.LoggerFactory
 import java.util.UUID
 
@@ -27,7 +28,7 @@ class CreateOrUpdateMissionWithActionsAndAttachedReporting(
 
     @Throws(IllegalArgumentException::class)
     fun execute(
-        // @UseCaseValidation<MissionEntity>(validator = MissionValidator::class)
+        @UseCaseValidation<MissionEntity>(validator = MissionWithEnvActionsValidator::class)
         mission: MissionEntity,
         attachedReportingIds: List<Int>,
         envActionsAttachedToReportingIds: List<EnvActionAttachedToReportingIds>,
